Main/Lambda/Batch_layer/spark_job_requirements.txt
python-dotenv
pandas
xgboost 